+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - Manually requesting code suggestions causing a SWT thread error ([!179](https://gitlab.com/gitlab-org/editor-extensions/gitlab-eclipse-plugin/-/merge_requests/179))
+- Always use caret offset for code suggestions ([!181](https://gitlab.com/gitlab-org/editor-extensions/gitlab-eclipse-plugin/-/merge_requests/181))
 
 ## 0.4.1 (2025-03-25)